rename method and move everything using server imports to server dir
